create enum for column indexes, move RemoveLastExpense to SpreadsheetService, and centralize startRow and searchColumn in a single class
