+++ BudgetAutomation/ExpenseLoggerApi/Constants/SpreadsheetDefaults.cs
@@ -1,0 +1,8 @@
+﻿namespace ExpenseLoggerApi.Constants
+{
+    public static class SpreadsheetDefaults
+    {
+        public const int StartRow = 15;
+        public const string SearchColumn = "B";
+    }
+}
