+++ BudgetAutomation/ExpenseLoggerApi/ExpenseLoggerApi.csproj
@@ -12,27 +12,27 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Amazon.Lambda.Core" Version="2.5.1"/>
-        <PackageReference Include="Amazon.Lambda.APIGatewayEvents" Version="2.7.1"/>
-        <PackageReference Include="Amazon.Lambda.Annotations" Version="1.6.1"/>
-        <PackageReference Include="Amazon.Extensions.Configuration.SystemsManager" Version="7.0.0"/>
-        <PackageReference Include="Amazon.Lambda.AspNetCoreServer.Hosting" Version="1.8.0"/>
-        <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.4"/>
-        <PackageReference Include="Google.Apis.Sheets.v4" Version="1.69.0.3694"/>
-        <PackageReference Include="Microsoft.AspNetCore.RateLimiting" Version="7.0.0-rc.2.22476.2"/>
+        <PackageReference Include="Amazon.Lambda.Core" Version="2.5.1" />
+        <PackageReference Include="Amazon.Lambda.APIGatewayEvents" Version="2.7.1" />
+        <PackageReference Include="Amazon.Lambda.Annotations" Version="1.6.1" />
+        <PackageReference Include="Amazon.Extensions.Configuration.SystemsManager" Version="7.0.0" />
+        <PackageReference Include="Amazon.Lambda.AspNetCoreServer.Hosting" Version="1.8.0" />
+        <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.4" />
+        <PackageReference Include="Google.Apis.Sheets.v4" Version="1.69.0.3694" />
+        <PackageReference Include="Microsoft.AspNetCore.RateLimiting" Version="7.0.0-rc.2.22476.2" />
     </ItemGroup>
 
     <ItemGroup>
-        <TrimmerRootAssembly Include="Newtonsoft.Json"/>
+        <TrimmerRootAssembly Include="Newtonsoft.Json" />
     </ItemGroup>
 
     <ItemGroup>
-        <InternalsVisibleTo Include="ExpenseLoggerApi.Tests"/>
+        <InternalsVisibleTo Include="ExpenseLoggerApi.Tests" />
     </ItemGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\SharedLibrary.Lambda\SharedLibrary.Lambda.csproj" />
-        <ProjectReference Include="..\SharedLibrary\SharedLibrary.csproj"/>
+        <ProjectReference Include="..\SharedLibrary\SharedLibrary.csproj" />
     </ItemGroup>
 
 </Project>
