+++ BudgetAutomation/ExpenseLoggerApi/Extension/ServiceCollectionExtensions.cs
@@ -64,7 +64,6 @@ public static class ServiceCollectionExtensions
 
         // Register Services
         services.AddScoped<SpreadsheetService>();
-        services.AddScoped<RemoveLoggerService>();
         services.AddSingleton<GoogleSheetsClientFactory>();
         services.AddSingleton<SheetsService>(sp =>
         {
