+++ BudgetAutomation/SharedLibrary/Enum/ExpenseColumnIndex.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SharedLibrary.Enum
+{
+    public enum ExpenseColumnIndex
+    {
+        Description = 0,
+        Amount = 7,
+        Category = 3
+    }
+}
